# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShiftTrack API",
    "version": "0.1.0",
    "description": "Rutas derivadas de los routers implementados. Contrato de comandos generado de Zod. Los payloads de módulos se validan en sus handlers; este inventario todavía no describe todos sus cuerpos."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "post__api_v1_auth_login",
        "summary": "POST /api/v1/auth/login",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "password": {
                    "type": "string",
                    "maxLength": 256,
                    "writeOnly": true
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "get__api_v1_auth_me",
        "summary": "GET /api/v1/auth/me",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "post__api_v1_auth_logout",
        "summary": "POST /api/v1/auth/logout",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/auth/change-password": {
      "post": {
        "operationId": "post__api_v1_auth_change_password",
        "summary": "POST /api/v1/auth/change-password",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/branches": {
      "get": {
        "operationId": "get__api_v1_branches",
        "summary": "GET /api/v1/branches",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      },
      "post": {
        "operationId": "post__api_v1_branches",
        "summary": "POST /api/v1/branches",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/devices/register": {
      "post": {
        "operationId": "post__api_v1_devices_register",
        "summary": "POST /api/v1/devices/register",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/devices/user-grants": {
      "post": {
        "operationId": "post__api_v1_devices_user_grants",
        "summary": "POST /api/v1/devices/user-grants",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/devices/challenge": {
      "post": {
        "operationId": "post__api_v1_devices_challenge",
        "summary": "POST /api/v1/devices/challenge",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/devices/session": {
      "post": {
        "operationId": "post__api_v1_devices_session",
        "summary": "POST /api/v1/devices/session",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/sync/push": {
      "post": {
        "operationId": "post__api_v1_sync_push",
        "summary": "POST /api/v1/sync/push",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "commands"
                ],
                "properties": {
                  "commands": {
                    "type": "array",
                    "maxItems": 20,
                    "items": {
                      "$ref": "#/components/schemas/Command"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/snapshot": {
      "get": {
        "operationId": "get__api_v1_snapshot",
        "summary": "GET /api/v1/snapshot",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/users": {
      "post": {
        "operationId": "post__api_v1_users",
        "summary": "POST /api/v1/users",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/users/{id}/deactivate": {
      "post": {
        "operationId": "post__api_v1_users__id__deactivate",
        "summary": "POST /api/v1/users/{id}/deactivate",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/clock/{action}": {
      "post": {
        "operationId": "post__api_v1_clock__action_",
        "summary": "POST /api/v1/clock/{action}",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/schedules": {
      "post": {
        "operationId": "post__api_v1_schedules",
        "summary": "POST /api/v1/schedules",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "firebaseIdToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token (no custom token)"
      },
      "deviceTransport": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Device TOKEN; solo transporte de pendientes y confirmaciones de equipo"
      }
    },
...
  }
}